decorator for User model
