+++ chapter5/tandibi/app/models/user.rb
@@ -66,21 +66,6 @@ class User < ApplicationRecord
 
   before_save :ensure_proper_name_case
 
-  def name
-    if last_name
-      "#{first_name} #{last_name}"
-    else
-      first_name
-    end
-  end
-
-  def profile_picture_url
-    @profile_picture_url ||= begin
-      hash = Digest::MD5.hexdigest(email)
-      "https://www.gravatar.com/avatar/#{hash}?d=wavatar"
-    end
-  end
-
   def to_param
     username
   end
